+++ views/new.ejs
@@ -11,18 +11,14 @@
     <input type="text" name="image" id="image">
     <br>
     <label for="stats">Stats: </label>
-    <label for="stats">attack </label>
+    <label for="attack">attack </label>
     <input type="number" name="attack" id="attack">
-    <label for="stats">defense </label>
+    <label for="defense">defense </label>
     <input type="number" name="defense" id="defense">
-    <label for="stats">speed </label>
+    <label for="speed">speed </label>
     <input type="number" name="speed" id="speed">
     <br>
     <input type="submit" value="Add new Pokemon">
 </form>
-
-
-
-
 
 <%- include('./partials/footer.ejs') %>
